refactor: move selected activioty store to context
