+++ packages/mobile/lib/contexts/dashboard/constants/drawer-component.constants.ts
@@ -1,0 +1,2 @@
+export const DRAWER_IN_ANIMATION_DURATION = 300
+export const DRAWER_OUT_ANIMATION_DURATION = 200
