+++ packages/mobile/lib/contexts/dashboard/stores/index.ts
@@ -1,1 +1,2 @@
 export * from './active-dashboard-tab.store'
+export * from './selected-activity.store'
