--- packages/mobile/lib/wallet/stores/index.ts
@@ -1,1 +1,0 @@
-export * from './selected-activity.store'
